+++ src/main/index.ts
@@ -7,6 +7,8 @@ export {BoxStyles} from './BoxStyles';
 
 export {HorizontalAlignmentEnum} from './HorizontalAlignmentEnum';
 
+export * as Sudoku from './sudoku/index';
+
 function drawBox(boxStyle: BoxStyle, data: string | string[][]) {
   new BoxDrawingBuilder().setData(data).setStyle(boxStyle).drawBox();
 }